Move getstate and setstate from Job to API
